+++ src/main/java/net/spell_engine/item/SpellEngineItems.java
@@ -8,16 +8,22 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
+import net.minecraft.registry.RegistryWrapper;
+import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.registry.tag.TagKey;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.Lazy;
 import net.spell_engine.SpellEngineMod;
+import net.spell_engine.api.spell.SpellTagsNumbered;
+import net.spell_engine.api.spell.Spell;
 import net.spell_engine.api.spell.registry.SpellRegistry;
 import net.spell_engine.compat.trinkets.SpellScrollTrinketItem;
 import net.spell_engine.compat.trinkets.TrinketsCompat;
 import net.spell_engine.fx.SpellEngineSounds;
 import net.spell_engine.spellbinding.SpellBinding;
 import net.spell_engine.spellbinding.SpellBindingBlock;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Comparator;
 
@@ -51,10 +57,22 @@ public class SpellEngineItems {
                     .sorted(Comparator.comparing(a -> a.getKey().get().getValue().getNamespace() + "_" + a.value().tier + "_" + a.getKey().get().getValue().getPath()))
                     .forEach((entry) -> {
                         var scroll = new ItemStack(SCROLL.get());
-                        if (ScrollItem.applySpell(scroll, entry, true)) {
+                        if (ScrollItem.applySpell(scroll, entry, resolveSpellPool(registryWrapper, entry))) {
                             content.add(scroll);
                         }
                     });
         });
+    }
+
+    @Nullable private static TagKey<Spell> resolveSpellPool(RegistryWrapper<Spell> wrapper, RegistryEntry<Spell> spellEntry) {
+        // Find the first tag in which spellEntry is contained
+        var tag = wrapper.streamTags()
+                .filter(t -> SpellTagsNumbered.isRegistered(t.getTagKey().get().id()) && t.contains(spellEntry))
+                .findFirst();
+        if (tag.isPresent()) {
+            return tag.get().getTagKey().get();
+        } else {
+            return null;
+        }
     }
 }